sw floattable, crashtesting: fix PDF export of fdo80989-1.docx
Converting the bugdoc to PDF crashed Writer layout since commit
ce3308a926f036b87515b8cd97d2b197063dc77a (tdf#61594 sw floattable:
import floating tables as split flys by default, 2023-04-12).

There were two problems here:

1) We assumed that when we move content to a next page, that page is
empty. This may not be true, and in case the next page has content,
we should insert the content at the start of the page, not at the end.
Specifying the sibling for MoveSubTree() fixes this.

2) We also hoped that if we ask for the next layout leaf, that will be
from the next page. Again, this may not be true,
SwFrame::GetNextSctLeaf() has quite complex conditions to detect this.
Split flys don't have to span over multiple columns, so just require
that the candidate is inside a different body then our anchor.

The bugdoc has to be loaded in hidden mode, otherwise we already calc
the layout on load, and then the problem is not visible.

Change-Id: I6e1533369db24c7c275ec1d7dceaddc4128c268a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150904
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins