sw floattable, crashtesting: fix PDF export of fdo72790-1.docx, part 1
Converting the bugdoc to PDF crashed Writer layout since commit
ce3308a926f036b87515b8cd97d2b197063dc77a (tdf#61594 sw floattable:
import floating tables as split flys by default, 2023-04-12).

What happens is that the DOCX file has a continuous section break, that
is mapped to a Writer section, but then the fly splitting code didn't
expect that the anchor might be inside a section frame.

Fix this by explicitly checking for an in-section anchor in
SwFrame::GetNextFlyLeaf(), which will give us a suitable insertion point
on the next page.

This does not fully fix the handling of the bugdoc, but at least now a
reduced sample doesn't crash and is handled correctly.

Change-Id: I52b5c901d899533bbb78131f443789d55db41000
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150956
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins