svx: rework the theme dialog to allow editing theme colors
In addition to editing of theme colors, this also changes the
way how theme colors are represented in the theme dialog and the
behaviour when changing applying a theme color change.

Previously a theme colors were applied with double-click, now a
double-click still applies, but automatically exits the dialog.
The selected theme is applied also when OK button is pressed.

Change-Id: Ic0f8399ede180d6ab0001a7f8b5dda0e7c49fa3c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150975
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>