tdf#86630 sw page number wizard: default to center of page
Defaulting to one side of the page only makes sense
if the footer is mirrored.
[Well, mirrored isn't the best word here
because it only refers to margin spacing.]

The only way to get mirrored footers is to have
separate left and right pages.

It would be nice to add mirroring to the wizard,
but that can be done separately.

First, lets just get a better default value in place.

Change-Id: I314c6a34414a696c6f2e27bb2b0747fdb86f3b22
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150998
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Paris Oplopoios <parisoplop@gmail.com>