tdf#154855 DOCX export: fix <wp:anchor layoutInCell="1"> with <wp:wrapNone>
The bugdoc has a table where shapes are anchored in cells. The wrap type
is set to "none" ("in front of text" in Word UI terms) and the follow
text flow ("layout in table ell" in Word UI terms) option is enabled.
The import into Writer is fine, but saving back to DOCX and checking in
Word is not: all shapes move to the left cells.

What happens is that since commit
e993638d5ecd33783f2eebdccfa87a81e5a8a2c5 (DOCX import: fix <wp:anchor
layoutInCell="1"> with <wp:wrapNone>, 2022-01-24), we ignore
layoutInCell in case wrapNone is used with a suitable horizontal
relation orient, since Word doesn't capture anchored shapes inside the
cells in practice in that case. This helps for the old bugdoc, but
introduces the problem that we now lose the layoutInCell
setting on export, which is not wanted.

Fix the problem by leaving the import and the layout unchanged, but do
the opposite mapping on export: if we have the relevant wrap and
horizontal relation orient case, then enable layoutInCell, even if the
doc model doesn't have it.

Note that Word itself behaves inconsistently here: if you change the
vertical offset of one of these shapes to position the shapes outside
the cell, Word does it, even if "layout in cell" is enabled.

Change-Id: I980a995d76afdee4716f7669081c688065427ee8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151000
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>