tdf#86630 sw page number wizard: set page style numbering type
Giving the user the opportunity to change the number type
is rather important. Well, maybe not too critical since I
assume that the default page number style is relatively accurate
for each locale, and at least for English we always just
use numbers for the page style.

However, a wizard is the perfect place to expose such a setting,
otherwise it is buried in page style settings - far from
where anyone would look when inserting a page number.

Change-Id: I9053504fa55d16ebf5424946b480cb044b79988c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151005
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>