Related: tdf#155017 Stop after the first pop in case the shell was not on stack
After commit 271f17a6022c64cbbe9befbc66a9eaaa0567d569 (tdf#155017: make
sure that the correct shell is popped from SfxDispatcher, 2023-04-26).

This restores the previous logic for a hypothetical case of a pop without
POP_UNTIL scheduled for a shell, which was not on the stack (so that it
wasn't removed in the code added in the commit mentioned above); I haven't
a reproducer, but it definitely can happen, given how things work here. So
I try to avoid breaking it even further :-)

Change-Id: Ied0a1406881fbefe3cf069f816c50da03a41dcd0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151034
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>