tdf#151715 sfx2: do not use FORCERELOAD for document stored in /tmp
This does not fix the actual bug, but is a workaround for a bug
somewhere in the Notebookbar code that was exposed by commit
5b5fb5e9c0580ecb70cf65882865bdaf16498fa5 "sfx2: do reload on SID_EDITDOC
in SfxViewFrame::ExecReload_Impl()".

For an ODF file, the Notebookbar is switched to editable by the call to
SetReadOnlyUI() from ReadOnlyUIGuard, but the XLS file takes a different
path and SetReadOnlyUI() is called in line 819.

Here the SfxViewFrame is not yet connected to the SfxObjectShell so
unsurprisingly it does not affect the UI; more surprising is that moving
the call to around line 859 leads to the SfxHintId::ModeChanged event
being handled in SfxViewFrame::Notify() with lots of things invalidated,
but the NotebookBar is entirely unimpressed by that.

Change-Id: I16f321bb3444eeae76348e20816af02938c3f46b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151089
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>