tdf#86630 sw page number wizard: mirror right/left
If the user puts the page numbers on the left or right
side of the page, they usually want that mirrored
on even and odd pages.

This got rather tricky, but in the end
I have enough safeguards that it seems
to work logically and stablely.
So I think it is ready to be submitted.

Change-Id: I321e575cd9f6718579ffee99f1258bffe26581f2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151152
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins