tdf#86630 sw page number wizard: mirror margins as well
This is only done in a very specific situation,
when no heading/footer is defined at all.
If we turn on mirroring at this point,
then be helpful and mirror the margins as well.

By default, this is irrelevant because the margins are identical.

It isn't necessary for page numbering,
but I think it is a nice touch to add something
that most people wouldn't find otherwise.
If they really want mirrorred page numbers,
then it is highly likely they also want mirrored margins.

Helping out people is the whole point of wizards after all.

Change-Id: I20cde65dc95ea3115951a4ea3a01534399f19f3d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151153
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>