tdf#155034 Revert "Update libxmlsec to 1.3.0"
This reverts commit 26bf26272bf525b59b4a4ce18b3ce14c1febfd7b.

Reason for revert: compiled version fails open/create some documents
on Ubuntu 20.04, which is still an Ubuntu-supported release.
Also fails to compile on 20.04 with built-in system NSS

The clinching reason is for running bibisects.
There was no compelling reason to make the change,
just routine maintenance. So if something breaks
or is annoying when doing routine maintenance, then revert it.

The previous version is still 1.2.37, released in Nov 2022.
So this will likely come up again relatively soon
if there is a security fix required.
But at least at the end of the 7.6 development cycle,
we can avoid the pain.

Change-Id: Ife387d6e4058b017ba18cba1fbcb2b2d50f52c12
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151118
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>