tdf#154211 SwNavigator: select range of content by shift + double-
click or shift + enter key

Enhancement to make a selection from the current cursor position in
the document to the position in the document that is navigated to
when a content entry in the Navigator tree is double clicked or the
enter key is pressed.

No selection is made when the current cursor position is in a table,
header, footer, footnote, or frame.

To make shift + double-click behavior for x11 and qt5 the same as
gtk3, FunctionSet::SetCursorAtPoint is added to the Shift case for
SelectionMode::Single in SelectionEngine::SelMouseButtonDown.

Change-Id: Id845dad8011ff7777a24f9b2730f10c62271c368
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151157
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>