sc drawstyles: Enable the clear DF command
.uno:StandardTextAttributes works for shape formatting already
(despite its name), but it's a different command than Format >
Clear DF. So let's connect the regular Clear DF command to the
same code.

Also includes a workaround for comments to maintain their size.
Changing the size (because of the "Fit height to text" setting)
is probably confusing for other shapes as well, but for comments
this requires an extra step to undo, as comments can't be resized
unless they're shown permanently. And manually resizing each
comment back will also result with inconsistency in their size.

Change-Id: I7c158523ecc4be18d17479632eaf9fd2a51dca12
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151161
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>