tdf#86630 sw page number wizard: add page total
1 - 3 is the format of the output.
Not totally clear, but it is at least not in English,
and I'm fairly certain I've seen this format used numerous times.

At least everything is there, and all the user needs to do is
modify the - to be whatever they want.

I checked to see if by some miracle localedata had a specification
for "X of Y" but it doesn't.

Change-Id: Iae1e74e612ec449f72086b3f5a5e32713fee4d27
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151173
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>