Simplify strtmpl.hxx
Replace bulky and difficult-to-read template argument names with
clearer ones; remove excessive whitespaces; reorder to avoid forward
declarations (these are remnants from the refactor from macros to
templates, when I tried to keep the old structure as close to the
original as possible).

And also a small code simplification in doubleToString.

Change-Id: I901cfb86daf410e85269538f36ebb6b553a4be76
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151125
Tested-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>