sw floattable, crashtesting: fix PDF export of forum-mso-en3-26783.docx
Converting the bugdoc to PDF crashed Writer layout since commit
ce3308a926f036b87515b8cd97d2b197063dc77a (tdf#61594 sw floattable:
import floating tables as split flys by default, 2023-04-12).

Inspecting what the layout does, the trouble was that the bugdoc has a
floating table, immediately followed by a non-floating table. This
breaks the invariant that floating tables are anchored in the next, text
node. In practice we anchored the floating table in the first cell of
the next table, which is not the same as a floating table followed by a
normal table. (For comparison, the DOC import simply lost the floating
table in this case.)

Fix the problem by making sure there is a text node after a floating
table, in case it's directly followed by a table (no text node
in-between).

Changed tests:

- testFloatingTablesAnchor: this still asserts that the flys are all
  anchored in the body text

- testTdf95755: the problematic table is still not as wide as it used to
  be in the bad case

Change-Id: I30b9ae423914b0b2cb1ebc7c1dce7efcd4cc3cf7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151251
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins