simplify SwSortUndoElement
No need for a union to save the minimal amount of memory. The code
is much harder to read an more error prone. Just duplicate source
and target - in one kace to be SwNodeOffset and other as OUString.

The assumption that SwNodeOffset doesn't have a default constr. is
also not true anymore, so we can just freely use that.

Change-Id: Ia96182eb2f872d0ca8f2e71ef2258d4b929c2d20
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151263
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>