tdf#137544 no need to broadcast in ~SdPage
It's not clear to me that we ever needed to do this - originally we did
not, and then we did, but only to avoid UserCall being triggered back
into us. Then the UserCall was explicitly cleared before the listener
were notified.
So just remove the notification again.
Which removes 10% of the overhead here.

Change-Id: I51f508e86c9785742bb092b7b3ae18f62f1f10f7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151266
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>