sw: fix crashtesting assert on tdf124604-1.docx
frmpaint.cxx:801: SwTextFrame::PaintSwFrame: Assertion `... || oTaggedParagraph || ...' failed.

Here the problem is that the numbering label on the last paragraph is so
indented that it ends up off the right end of the page, and so the loop
in SwTextPainter::DrawTextLine() terminates before seeing it.

Sigh, add another special case ...

(regression from commit 9b38beadf9eaf027b201cdf0ecb2bce5611014dd)

Change-Id: I6bb7385a10269d77d533c3b98462602e58a60317
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151299
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>