tdf#148956 sw: clear nAttrStart == paraEnd hints in DOCX
I don't think that native Writer will create a character
property at the end of a paragraph. (I couldn't do it manually
or by round-tripping a problematic DOCX file.)

However, DOCX does need to do this. However, a Ctrl-M
to clear direct formatting was not clearing out those
attributes (that don't span any text, but just sit
at the end of the node).

This patch fixes that problem.

Change-Id: I6bd93fc145d226f785287e6991b7ea6e3fcd1c9b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151306
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>