LOK: Navi-6 page selector SpinField
Grab the focus back to the spinfield in sidebar, so the document will
be able to scroll to the right page.
It seems there is a condition check, on move to page, that focus must be
not on document, or else it does not move.
AFAIK in lok code we dont try to simulate the focus of the core,
but maybe we should, for cases like this.

Change-Id: Ie9a30e6bd944cfa1729861eb0ac1924b946f4185
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151309
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
Tested-by: Szymon Kłos <szymon.klos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151701
Tested-by: Jenkins