tdf#136991 speed up rtf load
(*) since we are searching an ordered list, and we know that we want the
"next" range, a linear search is faster than a binary search, especially
when the list becomes large enough to exceed cache.
(*) Since element "i" and element  "i+1", we can start the second
search where the first search left off
(*) Since we are moving forward through the list, we can limit
subsequent searches by the results from the previous loop search.

This reduces the load time by 15%

Change-Id: I460339e3bc12e0a8bea39334173a34b2aa79515d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151323
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>