tdf#127322 svx: add back MID_GRAPHIC_URL to SvxBrush
This fixes the problem where gallery images
could not be set as page or paragraph backgrounds.

This URL attribute was removed in LO 6.1 with
author	Tomaž Vajngerl on 2018-03-09 09:41:44 +0100
commit c7ff07b2dab639b901c80f1f62efee4dd3ad549b
    remove MID_GRAPHIC_URL and related mappings, implementation

Most of this was restored shortly thereafter with
author	Tomaž Vajngerl on 2018-05-07 16:30:35 +0200
commit	7a8ed362eb163ac15a000ba1cfc74b58315800a1 (patch)
    [API CHANGE] revert and deprecate *BackGraphicURL add *BackGraphic

However, this piece wasn't restored.

Change-Id: Ia74d2562b61135f3294a7d6ebfa72cfadb4bdb5b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151342
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>