Provide std::u16string_view based o3tl::iterateCodePoints
...as requested in the comments of
<https://gerrit.libreoffice.org/c/core/+/151303> "a11y: Fix returning unpaired
surrogates when retrieving characters" (incl. the additional preAdjustIndex
parameter).

The type of the indexUtf16 parameter obviously needed to be adapted to
std::u16string_view's std::size_t.  But there is no obvious best choice for the
type of the incrementCodePoints parameter (int? std::ssize_t?), so lets leave it
as sal_Int32.

For simplicity of avoiding a Library_o3tl, and to allow o3tl::iterateCodePoints
to be used in the implementation of rtl_uString_iterateCodePoints now,
o3tl::iterateCodePoints is provided as an inline function defined in the include
file.

Change-Id: I8280ca11d2a943bd2b7150a266807b358f321a72
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151366
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>