tdf#152173 Don't display tooltip windows when application is inactive
Starting with macOS 13 Ventura, inactive applications receive mouse
move events so when LibreOffice is inactive, a mouse move event causes
a tooltip to be displayed. Since the tooltip window is attached to its
parent window (to ensure that the tooltip is above the parent window),
displaying a tooltip pulls the parent window in front of the windows
of all other inactive applications.

Change-Id: I35caf480a8afb42694957a4a6c2bec20be591ef3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151394
Tested-by: Jenkins
Reviewed-by: Patrick Luby <plubius@neooffice.org>
(cherry picked from commit 3bc7cc14706f47d740dfc5715970054922ca470c)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151373