tdf#86630 sw page number wizard: add to notebookbar insert menus
There are two places in the notebookbar where this
Insert page number could fit - in insert and in layout.

Both of these have "Title Page" in the menu at the right,
so this patch adds the page number wizard underneath that.
(Title pages should be setup first before page numbers are added.)

This is pretty non-controversial I would hope.
A follow-up patch will consider placing it on the main ribbon.

Change-Id: I195ee5a134a581cb6570485f84e4030b63d7895c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151396
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>