Revert "tdf#86630 sw page number wizard: add contextLabel"
This reverts 7.6 commit b89f65a5138fe74b7858c2817fae4204b021c1f5

and also take the opportunity to add a tooltip.

It was thought that the label would only be noticed
in tools - customize. However, I have discovered that it
also is shown in the notebookbar. Thus, it may very well
be used in other user-visible places.

Additionally, Heiko didn't like the term "wizard"
since it is a single-step dialog.

Change-Id: Ib11b68625eebf190d49ee8ce2b7b63378e248820
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151400
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>