sw floattable: fix UI for turning fly split off
Given a multi-page floating table of 2 pages, turning the "allow frame
to split across pages" checkbox off in the frame properties dialog
resulted in a 1 page document, but with 2 fly frames instead of just 1.

There were 3 problems here:

1) SwTabFrame::MakeAll() should join the follow table frames, similar to
what happens then the split bit on the table itself is turned off.  Note
that this is only safe if the frame is not chained, either.

2) SwTabFrame::Cut() should mark the follow fly for deletion. The check
here was broken, since we didn't do anything in case the split bit is
off for the fly frame. This looks OK, but we still need to handle the
case when it was on, but it's now off.

3) Finally SwFlyFrame::Format() always increases the height of flys to
at least MINFLY, so drop the HasArea() condition that prevented
deletion. Checking if the fly has content is probably enough.

With this, disabling the "split" checkbox on the frame dialog results in
a single fly frame, as expected.

For the testcase, improve Create1x2SplitFly() a bit, so it's closer to
what floattable.docx import would create (to help debugging) and also
improve SwFlyFrame::UpdateAttr_(), because previously that only worked
by accident, due to some rounding error on the UI. (So the fly size
changed even when that was not the intent.)

Change-Id: I7b33e6557282032c5d8ed2c3aea21b78a33bd6d3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151408
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins