loplugin:unnecessarygetstr
The plugin in its current form is probably not intending to find these cases.
But it happened to do so on macOS with libc++, where the std::basic_string ctors
taking a final defaulted Allocator argument are instead implemented by a pair of
overloaded ctors, one taking no Allocator at all (see
671d1c6cd14b28b5960ad56086299bd69533dfd8 "Adapt loplugin:unnecessarygetstr to
libc++").

But these changes here are useful regardless, so lets leave it at that.

Change-Id: I2776671e2953182bdcad36432951a75f82412ebb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151410
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>