Change text auto-fit alg. to also increase the scaling
When in edit mode, the text can be deleted, so the text box size
can become smaller, but the auto-fit algorithm didn't take into
account.
In this case we already have the font and spacing scaling already
set to a specific value and we need to find a scaling value where
the margin is the smallest.

This change also adds a test for the issue.

Change-Id: I6c52f06dfbf5a1e582f7b31aceabf4736498ee90
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151412
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 6c042848b688f64b3c56d65dd9dc5fe85412660a)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151411
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>