tdf#137972: when correcting PaMs, move them to the end
The cursor position is pushed in SwView::ExecSpellPopup, and popped
in the end, to restore the changed position e.g. when the action is
cancelled.

In the replacement, the PaM of the pushed cursor is updated in
DocumentContentOperationsManager::ReplaceRangeImpl, likely to avoid
several updates during the further processing. The stated goal is
to move all PaMs out of the deleted range, but the direction is not
stated explicitly.

If this correction is removed, the cursor PaM gets corrected anyway,
and arrives at the end of the replaced text. So, to fix the bug, and
to be compliant with the default case, let's correct to the end of
the range, instead of the random direction (GetPoint() may return a
position at either end, depending how PaM was created).

Change-Id: I83bcf01bcc0bf7277a9a34263b524b1212785814
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151462
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>