related tdf#138925 sw inputwin: get correct cell formula
The problem was that repeatedly pressing F2,
or any non-mouse movement between cells
was getting a "random" formula returned.

The only thing that worked was moving the mouse,
because on every pixel change, it re-loads
the RequestHelp tooltip that displays the formula,
thus loading it as the most current formula.

The example bug document easily crashes,
and everything about this is terribly imprecise,
so I'm not attempting to make any unit test here.

The code is modeled from GetContentAtPos,
which also does an unconditional const_cast.

Change-Id: Iecaf2e9a56bccef9a1e05bc0667caad6c3aeb109
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151550
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>