tdf#154933 Rename "Text Body" para style to "Body Text"
- Change only the UI name, leaving the programmatic name
which used by ODF and UNO API unchanged.

- DOCX shouldn't be affected either, as that style is
mapped to Word's "Text Body".

- RTF export used to use the UI name (at least in 7.5).
Which means that files created in older LO will show 2
styles for old and new names (same would happen also with
files created with a different UI language). It seems to
be no longer the case in current master.

Change-Id: I8a4866d5db5f964a5d45071fb09a99ed4996fae4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151653
Tested-by: Maxim Monastirsky <momonasmon@gmail.com>
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>