tdf#154775 sw layout: fix unexpected large table print area left margin
The bugdoc had 2 pages, but the second page's content was mostly a
table. This table was rendered outside the page frame, so it was
practically invisible.

This started after commit fd7749fddc5a767461dfced55369af48e5a6d561 (sw:
fix handling of table vs fly overlaps in the AddVerticalFlyOffsets case,
2020-02-14), and that behavior is still wanted to have correct table
borders, but that change introduced an inconsistent case: we started to
always accept an invalid frame area definition of the fly, but we only
shifted the content down in the text::HoriOrientation::NONE case.

Fix the problem by limiting the "accept invalid frame area definition"
based on the horizontal orientation, so we either both ignore the
validity flag & shift down or do none of this. This keeps the original
bug fixed, but addresses the new "no visible table" problem.

Note that this only fixes the regression part; once the image from the
header is removed and undo is invoked, a similar problem still happens.
But that is a pre-existing, separate problem.

(cherry picked from commit 1b5bc2ae2f19a190bf588a5a26c4d125c3960991)

Conflicts:
	sw/CppunitTest_sw_core_layout.mk
	sw/qa/core/layout/tabfrm.cxx
	sw/source/core/layout/tabfrm.cxx

Change-Id: Id752c169b7c7fb6b75e6603ad29aaafbb676b8a5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160395
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>