tdf#155235 Draw crashes on close: SIGSEGV in SalInstance:GetYieldMutex()
This is a regression from
    commit 3b7db802731826b6cc3b55100470b0c61c1f2dfa
    Author: Noel Grandin <noel.grandin@collabora.co.uk>
    Date:   Thu May 4 10:06:14 2023 +0200
    tdf#105404 [API CHANGE] add index to accessiblity change event

The list of children in DocumentFocusListener is getting out of sync
because something is not sending add/remove events for accessible
objects, a problem which was hidden before the commit above.

I can't find the add/remove problem, so do the next best thing - fix the
dispose methods of various accessible objects so that they clear
references to other objects, which results in the relevant objects being
destructed in an order which does not lead to crashes.

Change-Id: Iac7f48d6e3f2bdf30edb5c7fd3c5de861aec4b7a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151657
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>