tdf#155324 sw: layout: try not to MoveFwd onto a page created by page break
There is a ToX that is updated; it has about 4 pages worth of entries.

When the old entries are deleted, 2 of the pages become empty, and since
commit b9ef71476fd70bc13f50ebe80390e0730d1b7afb these pages are deleted.

While layouting the new entries, these are moved onto the page following
the ToX, which starts with a page break and contains lots of footnotes.

Now the footnotes reduce the space on the page available for the ToX
entries, and thus after CalcLayout() there are 5 ToX pages instead of 4.

Then the page numbers are inserted into the entries, and another layout
action deletes one of the ToX pages; now all the page numbers are too
large by 1.

Some ideas to fix this:
1) ignore a footnote when formatting a frame that is before the
   footnote anchor frame; similar to commit
   c79bf7865bff4e88cc201357370d8faeef8e6ad9
2) invalidate the last content on the page when moving forward the
   footnote container, similar to commit
   eb85de8e6b61fb3fcb6c03ae0145f7fe5478bccf; this doesn't look easy to
   do because as it turns out the footnote container is moved in
   SwLayoutFrame::Cut() 5 function calls inside MoveFwd() while the frame
   on which MoveFwd() is called is still on the page, so would probably
   somehow need to be detected in MoveFwd() itself?
3) don't move frames forward onto a page that was created by a page
   break - instead create a new frame.

Let's try 3) for now, only in SwFrame::GetNextSctLeaf().

(regression from commit b9ef71476fd70bc13f50ebe80390e0730d1b7afb)

Change-Id: I641f586799a5ddb4e2a6ff8e9de784e422ecc214
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151711
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>