sw floattable: fix UI / property dialog for shape containing table
The bugdoc contains a shape, containing a table (fly + draw format
pair). Trying to open the shape properties (position & size) dialog in a
debug build crashes.

We hit an assertion failure in SwTextFrame::MapModelToView(), because
the model position does not belong to the text node of the text frame.
This happens because SwFEShell::CalcBoundRect() uses GetCurrFrame(),
that normally points to the anchor frame for draw formats, but pointed
to a text frame inside the table (inside the shape) instead in the
problematic case.

Fix the problem by improving SwFEShell::SelectObj(), so in case it
selects a draw format instead of a fly format, then we don't use
SwCursorShell::SetCursor() (which works with a coordinate, and will find
the in-shape text nodes) but instead use SwCursorShell::SetSelection()
to explicitly set the cursor to the anchor position.

This went wrong in commit 7596e26fd259ce5445212949403e7cd32303b2bd (Add
SwTextBoxHelper::findShapes, 2014-06-24).

(cherry picked from commit 61f3c796702f725f2c65b53b79ab7e190d39b6b8)

Change-Id: Ifa0d2a245b97fc2ada247849e3f22579730a62e7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151824
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>