sw floattable: disable UI if the frame is chained already
TextFrames can be chained and we look at those fly frames as shapes. In
this case, we should not offer the "split fly" functionality, which
makes sense only when the fly frame is considered as a floating table.

The combination of these two features have no use-case and it would just
cause trouble when exporting to Word formats.

Fix the problem by making SwFramePage::Reset() more strict: it already
only allowed fly split for the case when the content was a single table,
but now it also requires no chaining.

The other way around (split fly -> no chain) still needs doing.

Change-Id: I06c240209a3a26519a5df953e22c9ee1b54aed34
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151822
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins