ScriptForge - database.GetRows() tdf#155204 error when no data
The complete expected bheviour is:
when there is no data returned by the query,
- either GetRows() returns an empty array, (Header := False)
- or GetRows() returns an array with a single
row containing the column names only (Header := True)

In the example given in the bug report,
GetRows() gives an unexpected error.

Actually the "end-of-file" status is tested
with the isAfterLast() indicator.
It seems preferable to rely on the Boolean value
returned by the first() and next() methods applied
on the resultset.

Change-Id: Ibe97dbbcb03d45ebb9184fab2733abe4e04963a6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151844
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
(cherry picked from commit d07cc6706ef3b382fa16a104c97b69bc2d2365e5)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151781
Tested-by: Xisco Fauli <xiscofauli@libreoffice.org>