tdf#155327 Detect default color in WordArt
It is possible to not set the character color explicitely in a WordArt
shape. In such case MS Office uses the scheme color 'tx1' from current
active scheme. The import of the character properties does not set this
color in the fill properties. The patch adds it, when the character fill
properties are converted to shape fill properties.

Change-Id: Ic72fd9f55bac1e2874cbf701ffa692ca4fbc9435
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151851
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>